requirements.txt
python-dotenv>=0.21.0
openai>=1.0.0
pydantic>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.7.0
requests>=2.25.0
